fix: ListItem entity: set quantity to Double from Integer and set defaults and feat: implement add /list-items endpoints with ListMembershipService and refactor

+++ src/main/java/hu/unideb/cartshare/model/dto/request/UpdateListItemRequestDto.java
@@ -14,7 +14,7 @@ public class UpdateListItemRequestDto {
 
     @NotNull(message = "Kitöltése kötelező")
     @Positive
-    private Integer quantity = 1;
+    private Double quantity = 1.0;
 
     @NotEmpty(message = "Kitöltése kötelező")
     private String unit;
